resolve linking error
A `Ppx_forest` module was referenced by both `forest_ppx` library and
the `ppx_forest` executable. This created a loop in the dependency
graph, under which the topological sorting is not defined, consequently
this lead to a compilation command with an incorrect ordering of
libraries (`forest_ppx` came before the `forest_parser`).